Polish source code
There are some changes in this commit.
pipeline.rs was made to combine multiple process
functions into single process when text are handled. However, the actual
text colorizing process was much simpler than I thought. So I decided to
remove it and replace it with LineColorizer.
And I remove TextColorizer because it is coded for more complex parsed token results. But It isnt' needed now.